third_party/blink/web_tests/platform/generic/http/tests/history/redirect-js-location-href-after-load-expected.txt

This page is the target of a redirect.

PASS: History item count should be 2 and is.


============== Back Forward List ==============
        http://127.0.0.1:8000/history/redirect-js-location-href-after-load.html
curr->  http://127.0.0.1:8000/history/resources/redirect-target.html#2
===============================================
